feat(cli): add --language flag; consolidate multilingual builder tests into the existing nemotron_speech_streaming test file
