+++ tests/e2e/models/nemotron_speech_streaming/test_nemotron_speech_streaming_builder.py
@@ -1,7 +1,12 @@
 # SPDX-FileCopyrightText: Copyright (c) 2026 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
 # SPDX-License-Identifier: Apache-2.0
 
-"""Synthetic load tests for the Nemotron Speech Streaming RNNT family."""
+"""Synthetic load tests for the Nemotron Speech Streaming RNNT family.
+
+Covers both checkpoints the family supports today:
+- ``nvidia/nemotron-speech-streaming-en-0.6b`` (monolingual).
+- ``nvidia/nemotron-3.5-asr-streaming-0.6b`` (multilingual, prompt_kernel MLP).
+"""
 
 from __future__ import annotations
 
@@ -30,17 +35,19 @@ class TestNemotronSpeechStreamingPlugin:
     MEL_BINS, CONV_KERNEL, SUB_CH = 8, 3, 4
     PRED_HIDDEN, PRED_LAYERS, JOINT_HIDDEN = 12, 2, 10
 
+    # Multilingual-only constants.
+    NUM_PROMPTS, PK_HIDDEN = 8, 32
+    STREAMING_RIGHT_CONTEXTS_SORTED_DESC = [13, 6, 3, 0]
+    STREAMING_CACHE_LEFT = 56
+    PROMPT_DICTIONARY = {"en-US": 0, "es-ES": 1, "de-DE": 2, "fr-FR": 3}
+
     @staticmethod
     def _write_config(model_dir, config):
         (model_dir / "config.json").write_text(json.dumps(config))
 
-    def test_load_weights_keys(self, tmp_path):
-        try:
-            import torch
-        except ImportError:
-            pytest.skip("torch required for synthetic NeMo archive test")
-
-        from tensorrt_model_connect.families.nemotron_speech_streaming import plugin
+    def _build_state_dict(self, with_prompt_kernel: bool = False):
+        """Build a synthetic state dict for encoder + predictor + joint (+ prompt_kernel)."""
+        import torch
 
         sd = make_nemo_state_dict(
             self.VOCAB,
@@ -58,7 +65,6 @@ class TestNemotronSpeechStreamingPlugin:
             sd.pop(f"encoder.layers.{i}.conv.batch_norm.running_mean", None)
             sd.pop(f"encoder.layers.{i}.conv.batch_norm.running_var", None)
 
-        # RNNT predictor: embedding includes the blank row at VOCAB.
         sd["decoder.prediction.embed.weight"] = torch.randn(self.VOCAB + 1, self.PRED_HIDDEN)
         for i in range(self.PRED_LAYERS):
             sd[f"decoder.prediction.dec_rnn.weight_ih_l{i}"] = torch.randn(
@@ -75,7 +81,16 @@ class TestNemotronSpeechStreamingPlugin:
         sd["joint.joint_net.1.weight"] = torch.randn(self.VOCAB + 1, self.JOINT_HIDDEN)
         sd["joint.joint_net.1.bias"] = torch.randn(self.VOCAB + 1)
 
-        nemo_cfg = {
+        if with_prompt_kernel:
+            pk_input_dim = self.HIDDEN + self.NUM_PROMPTS
+            sd["prompt_kernel.0.weight"] = torch.randn(self.PK_HIDDEN, pk_input_dim)
+            sd["prompt_kernel.0.bias"] = torch.randn(self.PK_HIDDEN)
+            sd["prompt_kernel.2.weight"] = torch.randn(self.HIDDEN, self.PK_HIDDEN)
+            sd["prompt_kernel.2.bias"] = torch.randn(self.HIDDEN)
+        return sd
+
+    def _build_nemo_cfg(self, att_context_size=None, prompt_dictionary=None):
+        cfg = {
             "target": "nemo.collections.asr.models.EncDecRNNTBPEModel",
             "model_defaults": {
                 "enc_hidden": self.HIDDEN,
@@ -111,16 +126,35 @@ class TestNemotronSpeechStreamingPlugin:
             },
             "preprocessor": {"features": self.MEL_BINS},
         }
+        if att_context_size is not None:
+            cfg["encoder"]["att_context_size"] = att_context_size
+        if prompt_dictionary is not None:
+            cfg["train_ds"] = {"prompt_dictionary": dict(prompt_dictionary)}
+        return cfg
 
-        make_nemo_archive(tmp_path, sd, nemo_cfg)
+    def _write_top_level_config(self, tmp_path, model_type: str):
         self._write_config(tmp_path, {
-            "model_type": "nemotron_speech_streaming",
+            "model_type": model_type,
             "hidden_size": self.PRED_HIDDEN,
             "num_hidden_layers": self.PRED_LAYERS,
             "num_attention_heads": 1,
             "vocab_size": self.VOCAB + 1,
             "mel_length": 80,
         })
+
+    def test_load_weights_keys(self, tmp_path):
+        """Monolingual en-0.6b: encoder + predictor + joint, no prompt_kernel."""
+        try:
+            import torch  # noqa: F401
+        except ImportError:
+            pytest.skip("torch required for synthetic NeMo archive test")
+
+        from tensorrt_model_connect.families.nemotron_speech_streaming import plugin
+
+        sd = self._build_state_dict(with_prompt_kernel=False)
+        nemo_cfg = self._build_nemo_cfg()
+        make_nemo_archive(tmp_path, sd, nemo_cfg)
+        self._write_top_level_config(tmp_path, "nemotron_speech_streaming")
 
         cfg = ModelConfig.from_dir(tmp_path)
         weights = plugin.load_weights(str(tmp_path), cfg)
@@ -137,8 +171,73 @@ class TestNemotronSpeechStreamingPlugin:
         assert weights["joint_enc_w"].shape == (self.HIDDEN, self.JOINT_HIDDEN)
         assert weights["joint_pred_w"].shape == (self.PRED_HIDDEN, self.JOINT_HIDDEN)
         assert weights["joint_out_w"].shape == (self.JOINT_HIDDEN, self.VOCAB + 1)
+        assert weights["_has_prompt_kernel"] is False
 
         overrides = plugin.get_bundle_config_overrides(cfg)
         assert overrides["runtime_strategy"] == "nemotron_speech_streaming_speech_to_text_rnnt"
         assert overrides["rnnt_blank_id"] == self.VOCAB
         assert overrides["rnnt_causal_downsampling"] is True
+        assert overrides["rnnt_has_prompt_kernel"] is False
+
+    def test_load_weights_multilingual_artifacts(self, tmp_path):
+        """Multilingual 3.5: prompt_kernel + 4 right-contexts + cache_left=56."""
+        try:
+            import torch  # noqa: F401
+        except ImportError:
+            pytest.skip("torch required for synthetic NeMo archive test")
+
+        from tensorrt_model_connect.families.nemotron_speech_streaming import plugin
+
+        sd = self._build_state_dict(with_prompt_kernel=True)
+        nemo_cfg = self._build_nemo_cfg(
+            att_context_size=[[self.STREAMING_CACHE_LEFT, r]
+                              for r in self.STREAMING_RIGHT_CONTEXTS_SORTED_DESC],
+            prompt_dictionary=self.PROMPT_DICTIONARY,
+        )
+        make_nemo_archive(tmp_path, sd, nemo_cfg)
+        self._write_top_level_config(tmp_path, "nemotron_3_5_asr_streaming")
+
+        cfg = ModelConfig.from_dir(tmp_path)
+        weights = plugin.load_weights(str(tmp_path), cfg)
+
+        assert weights["_streaming_right_contexts"] == self.STREAMING_RIGHT_CONTEXTS_SORTED_DESC
+        assert 1 not in weights["_streaming_right_contexts"]
+        assert weights["_streaming_cache_left"] == self.STREAMING_CACHE_LEFT
+
+        pk_input_dim = self.HIDDEN + self.NUM_PROMPTS
+        assert weights["_has_prompt_kernel"] is True
+        assert weights["_num_prompts"] == self.NUM_PROMPTS
+        assert weights["_pk_input_dim"] == pk_input_dim
+        assert weights["_pk_output_dim"] == self.HIDDEN
+        assert weights["pk_w0"].shape == (pk_input_dim, self.PK_HIDDEN)
+        assert weights["pk_w2"].shape == (self.PK_HIDDEN, self.HIDDEN)
+        assert weights["_prompt_dictionary"] == self.PROMPT_DICTIONARY
+
+        overrides = plugin.get_bundle_config_overrides(cfg)
+        assert overrides["rnnt_streaming_cache_left"] == self.STREAMING_CACHE_LEFT
+        assert overrides["rnnt_streaming_right_contexts"] == \
+            self.STREAMING_RIGHT_CONTEXTS_SORTED_DESC
+        assert overrides["rnnt_has_prompt_kernel"] is True
+        assert overrides["rnnt_num_prompts"] == self.NUM_PROMPTS
+        assert overrides["rnnt_prompt_dictionary"] == self.PROMPT_DICTIONARY
+
+    def test_missing_prompt_dictionary_raises(self, tmp_path):
+        """prompt_kernel without train_ds.prompt_dictionary must fail loudly."""
+        try:
+            import torch  # noqa: F401
+        except ImportError:
+            pytest.skip("torch required for synthetic NeMo archive test")
+
+        from tensorrt_model_connect.families.nemotron_speech_streaming import plugin
+
+        sd = self._build_state_dict(with_prompt_kernel=True)
+        nemo_cfg = self._build_nemo_cfg(
+            att_context_size=[[self.STREAMING_CACHE_LEFT, 13]],
+            prompt_dictionary=None,
+        )
+        make_nemo_archive(tmp_path, sd, nemo_cfg)
+        self._write_top_level_config(tmp_path, "nemotron_3_5_asr_streaming")
+
+        cfg = ModelConfig.from_dir(tmp_path)
+        with pytest.raises(ValueError, match="prompt_dictionary"):
+            plugin.load_weights(str(tmp_path), cfg)
